move files to new structure
