+++ src/app/app.component.ts
@@ -1,19 +1,18 @@
-import {Component} from 'angular2/core';
+import {Component,OnInit} from 'angular2/core';
 import {InitializeService} from './shared/services/initialize.service';
-import {Store} from '../store/store';
-import {LoginComponent} from './components/login/login.component';
-import {StartpageComponent} from './components/startpage/startpage.component';
-import {ManageComponent} from './components/manage/manage.component';
-import {ActivitiesComponent} from './components/activities/activities.component';
-import {AdministrationComponent} from './components/administration/administration.component';
-import {OnInit} from 'angular2/core';
+import {Store} from './store/store';
+import {LoginComponent} from './features/login/login.component';
+import {StartpageComponent} from './features/startpage/startpage.component';
+import {ManageComponent} from './features/manage/manage.component';
+import {ActivitiesComponent} from './features/activities/activities.component';
+import {AdministrationComponent} from './features/administration/administration.component';
 import {LoginService} from './shared/stubs/services/login.service';
 import {UserLoginDto} from './shared/stubs/dtos/user-login-dto';
-import {userIsAuthenticated} from '../store/actions/session.actions';
-import {userWantsToLogin} from '../store/actions/session.actions';
+import {userIsAuthenticated} from './store/actions/session.actions';
+import {userWantsToLogin} from './store/actions/session.actions';
 import {TenantLoginDto} from './shared/stubs/dtos/tenant-login-dto';
-import {activeTenantsOfUserLoaded} from '../store/actions/session.actions';
-import {backendCallFails} from '../store/actions/app.actions';
+import {activeTenantsOfUserLoaded} from './store/actions/session.actions';
+import {backendCallFails} from './store/actions/app.actions';
 import {ApplicationHeader} from './shared/components/application-header/application-header.component';
 
 @Component({
