+++ src/app/shared/components/application-header/application-header.component.ts
@@ -1,8 +1,8 @@
 import {Component} from 'angular2/core';
 import {Navigation} from '../../ui-elements/navigation';
-import {Store} from '../../../../store/store';
-import {logoutUser} from '../../../../store/actions/session.actions';
-import {mainNavigationItemClicked, secondaryNavigationItemClicked} from '../../../../store/actions/navigation.actions';
+import {Store} from '../../../store/store';
+import {logoutUser} from '../../../store/actions/session.actions';
+import {mainNavigationItemClicked, secondaryNavigationItemClicked} from '../../../store/actions/navigation.actions';
 import {LoginService} from '../../stubs/services/login.service';
 
 @Component({
