+++ src/app/store/reducers/assignments.reducer.ts
