+++ src/app/store/reducers/data.reducer.ts
