+++ src/app/store/stores/assignments.store.ts
