+++ src/app/store/stores/data.store.ts
