+++ src/app/store/stores/data/user-session.store.ts
@@ -1,5 +1,5 @@
-import {TenantLoginDto} from '../../../app/shared/stubs/dtos/tenant-login-dto';
-import {UserLoginDto} from '../../../app/shared/stubs/dtos/user-login-dto';
+import {TenantLoginDto} from '../../../shared/stubs/dtos/tenant-login-dto';
+import {UserLoginDto} from '../../../shared/stubs/dtos/user-login-dto';
 
 export interface IUserSessionStore {
   user:UserLoginDto;
