+++ src/main.ts
@@ -1,6 +1,6 @@
 import {bootstrap} from 'angular2/platform/browser';
 import {AppComponent} from './app/app.component';
-import {Store} from './store/store';
+import {Store} from './app/store/store';
 
 import {HTTP_PROVIDERS} from 'angular2/http';
 import {BACKEND_PROVIDERS} from './app/shared/stubs/services';
